+++ entry/src/main/cpp/core/formatter.h
@@ -1,0 +1,4 @@
+#pragma once
+#include <string>
+
+std::string formatLargeIntegerToScientific(const std::string& intStr);
