+++ entry/src/main/cpp/core/string_utils.h
@@ -1,0 +1,4 @@
+#pragma once
+#include <string>
+
+void replaceAll(std::string& str, const std::string& from, const std::string& to);
